integrate authorizedUser into user
