--- client/Redux/actions/authorizeUser.js
@@ -1,17 +1,0 @@
-export const login = (user) => {
-  return (dispatch) => {
-    dispatch({
-      type: 'authorized',
-      payload: user
-    })
-  }
-}
-
-export const logout = () => {
-  return (dispatch) => {
-    dispatch({
-      type: 'unauthorized',
-      payload: user
-    })
-  }
-}
