+++ client/Redux/actions/index.js
@@ -22,3 +22,21 @@ export const updateUser = (userData) => {
     payload: userData
   }
 }
+
+export const login = (user) => {
+  return (dispatch) => {
+    dispatch({
+      type: 'authorized',
+      payload: user
+    })
+  }
+}
+
+export const logout = () => {
+  return (dispatch) => {
+    dispatch({
+      type: 'unauthorized',
+      payload: user
+    })
+  }
+}
