+++ client/Redux/reducers/index.js
@@ -1,11 +1,9 @@
 import { combineReducers } from 'redux';
 import uploadProfilePictureReducer from './uploadProfilePictureReducer';
-import userAuthReducer from './userAuthReducer';
 import userReducer from './user';
 
 const reducers = combineReducers({
   uploadProfilePictureClickCount: uploadProfilePictureReducer,
-  authorizedUser: userAuthReducer,
   user: userReducer
 });
