+++ client/Redux/reducers/user.js
@@ -1,7 +1,21 @@
 const userReducer = (state = {}, action) => {
   switch (action.type) {
     case 'UPDATE':
-      return action.payload
+      return {
+        ...state,
+        posts: action.payload.posts,
+        userInfo: action.payload.userInfo
+      }
+    case 'authorized':
+      return {
+        ...state,
+        username: action.payload
+      };
+    case 'unauthorized':
+      return {
+        ...state,
+        username: null
+      };
     default:
       return state
   }
